+++ xa-exp04/xa-exp04/src/engine/rendering/glfw/GLFWView.cpp
@@ -73,5 +73,5 @@ std::tuple<int, int> GLFWView::getCursorPosition() {
 
     glfwGetCursorPos(window, &xpos, &ypos);
 
-    return { xpos, ypos };
+    return { (int) xpos, (int) ypos };
 }
